New mechanics and mana display.
For new mechanics we have Antimana. Antimana reduces your maximum mana and also slows mana regeneration.

There's also a HUD element showing your level of mana. This is basically the HPticker nicked from standard HD and made to respect Mana instead of health. The specifics of the mana display will be changed as I get a better hold on the HUD code (and have a good idea for what it might look like).