Some new things, one of which is incredibly stupid.
The book now gives an innacurate cound on your mana, and only updates the count every second.
The plasma fist now shoots 10 projectiles in a randomized cloud. Unfortuantely, the pitch input for A-FireProjectile is also interpreted as setting the value of pitch, and not just reading from it. To get around it I've saved pitch as ptch, and replace the buggered pitch with ptch after it's done. There must be a less stupid way to do this.
Changed the plasma animation just because I can.
Also changed the zscipt version to 3.3.2 to match the target version.